Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    float radius, area;
+    double radius, area;
     scanf("%.2lf",&radius);
     area = (3.14*num*num);
     printf("Area: %.2lf\n", area);
